Move browserWithExperimentAPIs to separate file
